+++ src/main/java/com/peeko32213/unusualprehistory/datagen/DataGenerators.java
@@ -9,6 +9,7 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.models.blockstates.BlockStateGenerator;
+import net.minecraftforge.common.data.DatapackBuiltinEntriesProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.data.event.GatherDataEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -63,6 +64,12 @@ public class DataGenerators {
         BlockTagsGenerator blockTagGenerator = generator.addProvider(evt.includeServer(),
                 new BlockTagsGenerator(packOutput, lookupProvider, helper));
         generator.addProvider(evt.includeServer(), new ItemTagsGenerator(packOutput, lookupProvider, blockTagGenerator.contentsGetter(), helper));
+
+        DatapackBuiltinEntriesProvider datapackProvider = new RegistryDataGenerator(packOutput, lookupProvider);
+        CompletableFuture<HolderLookup.Provider> customLookupProvider = datapackProvider.getRegistryProvider();
+        generator.addProvider(evt.includeServer(), datapackProvider);
+
+
     }
 
 }